# Speed up calculate_tree_depth()