This PR speeds up the node depth calculations, see below for some results pre and post change. As such, it helps to solve #34

```
library(randomForest)
library(ranger)
library(randomForestExplainer)

set.seed(12)

fit <- randomForest(Sepal.Width~., data = iris)
fit2 <- ranger(Sepal.Width~., data = iris)

system.time(  # 6.25 seconds -> 2.0 seconds
  out <- min_depth_distribution(fit)
)
head(out)
#   tree     variable minimal_depth
# 1    1 Petal.Length             0
# 2    1  Petal.Width             3
# 3    1 Sepal.Length             5
# 4    1      Species             3
# 5    2 Petal.Length             3
# 6    2  Petal.Width             1

system.time(  # 4.8 seconds -> 0.70 seconds
  out <- min_depth_distribution(fit2)
)
head(out)
#   tree     variable minimal_depth
# 1    1 Petal.Length             0
# 2    1  Petal.Width             1
# 3    1 Sepal.Length             1
# 4    1      Species             3
# 5    2 Petal.Length             1
# 6    2  Petal.Width             1
```

You can test it via

```
remotes::install_github("ModelOriented/randomForestExplainer", ref = remotes::github_pull(35))
```